+++ Sources/NSURL+Convertible.swift
@@ -2,17 +2,28 @@ import Foundation
 
 /**
  NSURL Convertible implementation
-
- This implementation assumes the given value is a string which can be used to create a NSURL
- if not a `MapperError` is thrown
  */
 extension NSURL: Convertible {
+    /**
+     Create a NSURL from Mapper
+
+     - parameter value: The object (or nil) passed from Mapper
+
+     - throws: MapperError.ConvertibleError if the passed value is not a String
+     - throws: MapperError.CustomError      if the passed value a String but the NSURL initializer returns nil
+
+     - returns: The created NSURL
+     */
     @warn_unused_result
     public static func fromMap(value: AnyObject?) throws -> NSURL {
-        if let string = value as? String, let URL = NSURL(string: string) {
+        guard let string = value as? String else {
+            throw MapperError.ConvertibleError(value: value, type: String.self)
+        }
+
+        if let URL = NSURL(string: string) {
             return URL
         }
 
-        throw MapperError()
+        throw MapperError.CustomError(field: nil, message: "'\(string)' is not a valid NSURL")
     }
 }